client: Remove code examples from comments
Example with RCP left in the code, because it can't be properly represented with examples.
